Use quotes for {{t}} argument everywhere
Previous to Ember 1.0.0-rc.8, `{{foo bar}}` and `{{foo "bar"}}` were the
same. In 1.0.0-rc.8, `{{foo bar}}` will call `foo` with the value
of `currentContext.bar`, whereas `{{foo "bar"}}` will call `foo` with
the literal `"bar"`. Ember now takes care of the binding to `bar` in the
unquoted case. Thus, in the documentation and specs, I want to use the
quoted version since they're using names of translation keys, not
pointers to names of translation keys.