Update README to avoid confusion
Update README.md with a small note in the 'Requirements' section, to avoid confusion and a common error. 

Simply referencing the lightweight runtime version of handlebars (`handlebars.runtime.js`) in your project is not enough, the runtime version doesn't contain the compile function required to compile the templates. The full library should be included. 

The fact that you should use the full library isn't noted anywhere. I'm sure this will help people installing the plugin avoid the common error. 

Related to #45 and #63.