Fix TranslateableAttributes tests on Ember Canary
These tests included
```js
this.renderTemplate('{{view titleTranslation="foo.bar"}}');
```
which breaks on Ember Canary because the `{{view}}` helper no
longer assumes `Ember.View` by default, but instead asks the
container for `view:default`.

This commit changes `renderTemplate` to create an `Ember.Container`,
register an anonymous subclass of `Ember.View` as `view:toplevel`,
and set that container as the container for the created view.