{{t}} helper should update multiple bound properties.
The first observed property triggers a rerender, effectively removing it
from the DOM until the render loop runs.  The second observer examines
the view, finds that view.$() returns null, and assumes it's safe to
destroy itself.

To work around the above, keep the observer around when the view is in
"preRender" state.