Dynamic keys are not being translated
Things like `{{t dynamicKey}}` are not working ATM. This PR includes a
failing test.

The {{t}} helper (signature t(params, hash, options, env)) is receiving
a an array if `Stream` objects in `params`, not the actual values coming
from that stream.

I suspect (because I saw this in the past) that is because of the way
the helper is being registered. The registration of helpers has changed
a few times since HTMLBars.