remove Stream awareness from service:i18n
The only time that the `service:i18n` needs to expose a `localeStream` is when
we're using the legacy helper (Ember 1.12). Therefore, instead of always having
the service build that Stream, we use the initializer to set it up only when
we're installing the legacy helper. This makes the library completely streamless
on Ember 1.13+.