Use public API for container / registry functions.
As of Ember 2.3, using `this.container` from within an instance that is
looked up from the container will issue a deprecation warning suggesting
that you should refactor away from the private `this.container` usage to
the new public API for using container/registry features
(`Ember.getOwner(this)`).

This change pulls in the `ember-getowner-polyfill` to allow us to use
`getOwner` on all Ember versions (tested back to 1.10 in that addon's
CI).